Rename inline-condition-model file to condition-model
